# Avoid expesive pointer copy in capi nodegroup