@enxebre 2024-05-09
some benchMarks
before
```
➜  cluster-autoscaler git:(release-4.16) ✗ go test -v -bench=. -count 5 -run=NodeGroupsBenchmark ./cloudprovider/clusterapi/...
goos: darwin
goarch: amd64
pkg: k8s.io/autoscaler/cluster-autoscaler/cloudprovider/clusterapi
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
BenchmarkNodeGroups
I0509 16:35:53.758484   50802 clusterapi_controller.go:426] Using version "v1alpha3" for API group "cluster.x-k8s.io"
I0509 16:35:53.758676   50802 clusterapi_controller.go:541] Resource "machinedeployments" available
I0509 16:35:53.758740   50802 clusterapi_controller.go:541] Resource "machinedeployments" available
I0509 16:35:53.758749   50802 clusterapi_controller.go:541] Resource "machinesets" available
I0509 16:35:53.758754   50802 clusterapi_controller.go:541] Resource "machines" available
I0509 16:35:53.758760   50802 clusterapi_controller.go:541] Resource "machinepools" available
BenchmarkNodeGroups/NodeGroups
BenchmarkNodeGroups/NodeGroups-12         	     798	   1417817 ns/op
BenchmarkNodeGroups/NodeGroups-12         	     814	   1447729 ns/op
BenchmarkNodeGroups/NodeGroups-12         	     711	   1454166 ns/op
BenchmarkNodeGroups/NodeGroups-12         	     777	   1422128 ns/op
BenchmarkNodeGroups/NodeGroups-12         	     788	   1486356 ns/op
PASS
ok  	k8s.io/autoscaler/cluster-autoscaler/cloudprovider/clusterapi	7.960s
```

After
```
➜  cluster-autoscaler git:(capi-nodegroups) ✗ go test -v -bench=. -count 5 -run=NodeGroupsBenchmark ./cloudprovider/clusterapi/...
goos: darwin
goarch: amd64
pkg: k8s.io/autoscaler/cluster-autoscaler/cloudprovider/clusterapi
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
BenchmarkNodeGroups
I0509 16:32:44.577941   50440 clusterapi_controller.go:426] Using version "v1alpha3" for API group "cluster.x-k8s.io"
I0509 16:32:44.578106   50440 clusterapi_controller.go:543] Resource "machinedeployments" available
I0509 16:32:44.578137   50440 clusterapi_controller.go:543] Resource "machinedeployments" available
I0509 16:32:44.578142   50440 clusterapi_controller.go:543] Resource "machinesets" available
I0509 16:32:44.578145   50440 clusterapi_controller.go:543] Resource "machines" available
I0509 16:32:44.578147   50440 clusterapi_controller.go:543] Resource "machinepools" available
BenchmarkNodeGroups/NodeGroups
BenchmarkNodeGroups/NodeGroups-12         	    1215	    948669 ns/op
BenchmarkNodeGroups/NodeGroups-12         	    1120	   1011799 ns/op
BenchmarkNodeGroups/NodeGroups-12         	    1159	   1052832 ns/op
BenchmarkNodeGroups/NodeGroups-12         	     999	   1023880 ns/op
BenchmarkNodeGroups/NodeGroups-12         	    1122	   1020194 ns/op
PASS
ok  	k8s.io/autoscaler/cluster-autoscaler/cloudprovider/clusterapi	7.701s
```